Struct underscore_util::symbol::Symbols

pub struct Symbols<V: Clone> {
    pub strings: Rc<SymbolMap<Symbol>>,
    pub table: HashMap<Symbol, Vec<V>>,
    // some fields omitted
}

A Scoped Map that takes any K and V

Fields
strings: Rc<SymbolMap<Symbol>> table: HashMap<Symbol, Vec<V>>
Methods

impl<V: Clone> Symbols<V>

pub fn new(strings: Rc<SymbolMap<Symbol>>) -> Self

A new Symbols Instance

pub fn begin_scope(&mut self)

Adds a new scope to the table

pub fn end_scope(&mut self)

Recursivly destorys the scopes

pub fn enter(&mut self, symbol: Symbol, data: V)

Enters a peice of data into the current scope

pub fn look(&self, symbol: Symbol) -> Option<&V>

Looks in the table for the Symbol and if found returns the top element in the stack of Vec

pub fn name(&self, symbol: Symbol) -> String

Finds the name given to a Symbol

pub fn symbol(&mut self, name: &str) -> Symbol

pub fn replace(&mut self, symbol: Symbol, data: V)

Inserts the Symbol into the table